Added ddk commands in cdk deploy
