+++ backend/dataall/cdkproxy/requirements.txt
@@ -15,3 +15,4 @@ jinja2==3.0.3
 werkzeug==2.0.2
 constructs>=10.0.0,<11.0.0
 aws_ddk_core==0.3.1
+git-remote-codecommit
